+++ tests/test_enterprise/test_utils.py
@@ -1,0 +1,41 @@
+# -*- coding: utf-8 -*-
+"""
+Tests for the `edx-enterprise` utils module.
+"""
+
+import unittest
+
+import ddt
+from pytest import mark
+
+from enterprise.utils import get_idiff_list
+
+
+@mark.django_db
+@ddt.ddt
+class TestGetDifferenceList(unittest.TestCase):
+    """
+    Tests for :method:`get_idiff_list`.
+    """
+
+    @ddt.unpack
+    @ddt.data(
+        (
+            [],
+            [],
+            []
+        ),
+        (
+            ['DUMMY1@example.com', 'dummy2@example.com', 'dummy3@example.com'],
+            ['dummy1@example.com', 'DUMMY3@EXAMPLE.COM'],
+            ['dummy2@example.com']
+        ),
+        (
+            ['dummy1@example.com', 'dummy2@example.com', 'dummy3@example.com'],
+            [],
+            ['dummy1@example.com', 'dummy2@example.com', 'dummy3@example.com'],
+        )
+    )
+    def test_get_idiff_list_method(self, all_emails, registered_emails, unregistered_emails):
+        emails = get_idiff_list(all_emails, registered_emails)
+        self.assertEqual(sorted(emails), sorted(unregistered_emails))
