+++ tests/test_utilities.py
@@ -13,6 +13,7 @@ from faker import Factory as FakerFactory
 from pytest import mark, raises
 
 from django.core import mail
+from django.core.exceptions import ValidationError
 from django.test import override_settings
 
 from enterprise import utils
@@ -21,10 +22,12 @@ from enterprise.models import (
     EnterpriseCustomerBrandingConfiguration,
     EnterpriseCustomerIdentityProvider,
     EnterpriseCustomerUser,
+    PendingEnterpriseCustomerUser,
 )
 from integrated_channels.sap_success_factors.models import SAPSuccessFactorsEnterpriseCustomerConfiguration
 from test_utils import TEST_UUID, create_items, fake_catalog_api
 from test_utils.factories import (
+    FAKER,
     EnterpriseCustomerFactory,
     EnterpriseCustomerIdentityProviderFactory,
     EnterpriseCustomerUserFactory,
@@ -1686,3 +1689,76 @@ class TestSAPSuccessFactorsUtils(unittest.TestCase):
         # audit course modes are not filtered out
         filtered_course_modes = utils.filter_audit_course_modes(self.customer, course_modes)
         assert len(filtered_course_modes) == 5
+
+
+@mark.django_db
+@ddt.ddt
+class TestValidateEmailToLink(unittest.TestCase):
+    """
+    Tests for :method:`validate_email_to_link`.
+    """
+
+    def test_validate_email_to_link_normal(self):
+        email = FAKER.email()  # pylint: disable=no-member
+        assert PendingEnterpriseCustomerUser.objects.filter(user_email=email).count() == 0, \
+            "Precondition check - should not have PendingEnterpriseCustomerUser"
+        assert EnterpriseCustomerUser.objects.count() == 0, \
+            "Precondition check - should not have EnterpriseCustomerUser"
+
+        exists = utils.validate_email_to_link(email)  # should not raise any Exceptions
+        assert exists is False
+
+    @ddt.unpack
+    @ddt.data(
+        ("something", "something", utils.ValidationMessages.INVALID_EMAIL),
+        ("something_again@", "something_else", utils.ValidationMessages.INVALID_EMAIL)
+    )
+    def test_validate_email_to_link_invalid_email(self, email, raw_email, msg_template):
+        assert EnterpriseCustomerUser.objects.get_link_by_email(email) is None, \
+            "Precondition check - should not have EnterpriseCustomerUser or PendingEnterpriseCustomerUser"
+
+        expected_message = msg_template.format(argument=raw_email)
+
+        with raises(ValidationError, match=expected_message):
+            utils.validate_email_to_link(email, raw_email)
+
+    @ddt.data(True, False)
+    def test_validate_email_to_link_existing_user_record(self, ignore_existing):
+        user = UserFactory()
+        email = user.email
+        existing_record = EnterpriseCustomerUserFactory(user_id=user.id)
+        assert not PendingEnterpriseCustomerUser.objects.exists(), \
+            "Precondition check - should not have PendingEnterpriseCustomerUser"
+        assert EnterpriseCustomerUser.objects.get(user_id=user.id) == existing_record, \
+            "Precondition check - should have EnterpriseCustomerUser"
+
+        if ignore_existing:
+            exists = utils.validate_email_to_link(email, ignore_existing=True)
+            assert exists
+        else:
+            expected_message = utils.ValidationMessages.USER_ALREADY_REGISTERED.format(
+                email=email, ec_name=existing_record.enterprise_customer.name
+            )
+
+            with raises(ValidationError, match=expected_message):
+                utils.validate_email_to_link(email)
+
+    @ddt.data(True, False)
+    def test_validate_email_to_link_existing_pending_record(self, ignore_existing):
+        email = FAKER.email()  # pylint: disable=no-member
+        existing_record = PendingEnterpriseCustomerUserFactory(user_email=email)
+        assert PendingEnterpriseCustomerUser.objects.get(user_email=email) == existing_record, \
+            "Precondition check - should have PendingEnterpriseCustomerUser"
+        assert not EnterpriseCustomerUser.objects.exists(), \
+            "Precondition check - should not have EnterpriseCustomerUser"
+
+        if ignore_existing:
+            exists = utils.validate_email_to_link(email, ignore_existing=True)
+            assert exists
+        else:
+            expected_message = utils.ValidationMessages.USER_ALREADY_REGISTERED.format(
+                email=email, ec_name=existing_record.enterprise_customer.name
+            )
+
+            with raises(ValidationError, match=expected_message):
+                exists = utils.validate_email_to_link(email)
